refactor: convert from monorepo to single-package
