+++ package.json
@@ -1,27 +1,44 @@
 {
-  "name": "elemental-monorepo",
-  "private": true,
-  "version": "1.0.0",
+  "name": "elemental-js",
+  "version": "0.9.0",
+  "description": "A tiny, lightweight, fast, and dependency-free vanilla JavaScript UI lib.",
   "author": "Adomas Stefanovičius",
   "license": "MIT",
   "type": "module",
+  "main": "./dist/elemental.umd.cjs",
+  "module": "./dist/elemental.js",
+  "exports": {
+    ".": {
+      "require": "./dist/elemental.umd.cjs",
+      "import": "./dist/elemental.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "keywords": [
+    "ui",
+    "framework",
+    "web",
+    "component",
+    "tiny",
+    "lightweight",
+    "lib",
+    "library",
+    "fast",
+    "vanilla",
+    "javascript",
+    "js"
+  ],
   "scripts": {
-    "dev": "pnpm --filter elemental-js dev",
-    "build": "pnpm --filter elemental-js build",
-    "dev:site": "pnpm --filter elemental-js build:watch & pnpm --filter elemental-js dev & pnpm --filter elemental-site dev & wait",
-    "build:site": "pnpm --filter elemental-site build",
-    "preview:site": "pnpm --filter elemental-site preview"
+    "dev": "vitest",
+    "build": "vite build"
   },
   "devDependencies": {
     "@changesets/cli": "^2.27.12",
+    "happy-dom": "^16.7.3",
     "prettier": "^3.4.2",
-    "vite": "^6.0.11"
-  },
-  "engines": {
-    "node": ">=20.0.0",
-    "pnpm": ">=9.0.0"
-  },
-  "workspaces": [
-    "packages/*"
-  ]
+    "vite": "^6.0.11",
+    "vitest": "^3.0.4"
+  }
 }
