--- packages/lib/README.md
@@ -1,54 +1,0 @@
-# Elemental.js
-
-Welcome to Elemental.js – the lightweight, no-nonsense JavaScript library for building reactive UIs with a touch of magic. Don't just write code - _control the elements_.
-
-## Introduction
-
-Tired of the heavyweight frameworks? Want something simple yet powerful? Elemental.js is your spellbook for creating dynamic, responsive web interfaces with minimal fuss. Simple syntax, powerful features - all in one compact package.
-
-## Features
-
-- **Effortless Reactivity**: Breathe life into your UI with a straightforward reactivity model. Watch as your UI updates magically as your data changes.
-- **Zero Dependencies**: With Elemental.js, what you see is what you get - no external dependencies, just pure JavaScript wizardry.
-- **Lightweight and Fast**: Designed to be lean and mean. Enjoy the speed without the bloat of larger frameworks.
-- **Easy to Learn**: Jump right in, even if you're not a sorcerer of JavaScript. Clear syntax makes it a breeze.
-
-## Getting Started
-
-Add some elemental power to your project in just a few lines:
-
-```javascript
-import { el } from 'elemental-js'
-
-// Create a simple element
-const greeting = el`div "Hello, Elemental World!"`
-```
-
-## Example of a Counter Component
-
-```javascript
-import { def, el } from 'elemental-js'
-
-function counter() {
-  const count = def(0)
-  const increment = () => ++count.val
-  return el`button onclick="${increment}" "Count is: ${count}"`
-}
-
-document.body.append(counter())
-```
-
-## Future Plans
-
-Elemental.js is just at the beginning of its journey, here are the plans for its evolution:
-
-- **Refactoring for Efficiency**: The first version was a thrilling rush of creation. Next, I'm rolling up my sleeves to refactor the code. This will not only make Elemental.js more efficient but also easier to understand and contribute to.
-- **Comprehensive Testing**: Testing is crucial, and it's on my list. I plan to implement a thorough testing process to ensure reliability and stability in every element controlled by Elemental.js.
-- **Documentation Development**: As the library matures, so will its documentation. Detailed guides and examples will be provided, making it easier for users to master Elemental.js and use it to its full potential.
-- **Community Contributions**: Once the foundation is solid, I'll be laying out clear guidelines for contributions. I'm excited to see how fellow developers can add their magic to Elemental.js, enhancing it with their unique ideas and improvements.
-- **Performance Optimization**: Continuous performance tuning is a given. Expect leaner, faster code in future releases, ensuring that Elemental.js remains a lightweight yet powerful tool for developers.
-- **Feature Expansion**: I have a vision of adding more features while keeping the core simple and intuitive. Stay tuned for updates that bring new capabilities to your elemental toolkit.
-
-## Let the Elements Be With You
-
-Embrace the power of Elemental.js and let the elements guide your path to creating enchanting user interfaces. Who knew controlling the elements could be this easy?
