--- packages/lib/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "elemental-js",
-  "version": "0.9.0",
-  "description": "A tiny, lightweight, fast, vanilla JavaScript UI lib.",
-  "author": "Adomas Stefanovičius",
-  "license": "MIT",
-  "type": "module",
-  "main": "./dist/elemental.umd.cjs",
-  "module": "./dist/elemental.js",
-  "exports": {
-    ".": {
-      "require": "./dist/elemental.umd.cjs",
-      "import": "./dist/elemental.js"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "keywords": [
-    "ui",
-    "framework",
-    "web",
-    "component",
-    "tiny",
-    "lightweight",
-    "lib",
-    "library",
-    "fast",
-    "vanilla",
-    "javascript",
-    "js"
-  ],
-  "scripts": {
-    "dev": "vitest",
-    "build": "vite build",
-    "build:watch": "vite build --watch"
-  },
-  "devDependencies": {
-    "happy-dom": "^16.7.3",
-    "vitest": "^3.0.4"
-  }
-}
