--- packages/site/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "elemental-site",
-  "private": true,
-  "version": "1.0.0",
-  "author": "Adomas Stefanovičius",
-  "license": "MIT",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "elemental-js": "workspace:*"
-  }
-}
