--- packages/site/src/examples/index.html
@@ -1,28 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <title>Elemental.JS</title>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="Cache-Control" content="no-store, no-cache, must-revalidate, max-age=0" />
-    <meta http-equiv="Pragma" content="no-cache" />
-    <meta http-equiv="Expires" content="0" />
-    <script type="module" src="index.js"></script>
-  </head>
-  <body>
-    <header>
-      <nav>
-        <a href="/"><span class="btn">Home</span><i data-icon="home"></i></a>
-        <a href="#"><span class="btn">Examples</span><i data-icon="flask"></i></a>
-        <a href="https://github.com/Stefanovicius/elemental-js" target="_blank">
-          <span class="btn">GitHub</span><i data-icon="github"></i>
-        </a>
-      </nav>
-    </header>
-    <main></main>
-    <footer>
-      Made with <span>❤️</span> by
-      <a href="https://github.com/Stefanovicius" target="_blank">Adomas Stefanovičius</a>
-    </footer>
-  </body>
-</html>
