--- packages/site/src/index.html
@@ -1,146 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <title>Elemental.js</title>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" href="./assets/favicon.png" type="image/png" sizes="182x182" />
-    <script type="module" src="./index.js"></script>
-  </head>
-  <body>
-    <header>
-      <nav>
-        <a href="#"><span class="btn">Home</span><i data-icon="home"></i></a>
-        <a href="examples/"><span class="btn">Examples</span><i data-icon="flask"></i></a>
-        <a href="https://github.com/Stefanovicius/elemental-js" target="_blank">
-          <span class="btn">GitHub</span><i data-icon="github"></i>
-        </a>
-      </nav>
-      <img id="wand" src="./assets/wand.webp" alt="Magic Wand" width="400" height="400" />
-      <h1>Elemental.js</h1>
-      <p class="tagline">
-        The lightweight, no-nonsense JavaScript library for building reactive UIs with a touch of
-        magic. Don't just write code - control the elements!
-      </p>
-    </header>
-    <main>
-      <section>
-        <h2 id="intro">
-          <span style="color: inherit; text-decoration: none; cursor: pointer">Intro</span>
-        </h2>
-        <p id="scroll-target">
-          Tired of the heavyweight frameworks? Want something simple yet powerful? Elemental.js is
-          your spell-book for creating dynamic, responsive web interfaces with minimal fuss. Simple
-          syntax, powerful features - all in one compact package, with no external dependencies.
-        </p>
-      </section>
-      <section>
-        <h2>Features</h2>
-        <ul>
-          <li>
-            <strong>⚡ Effortless Reactivity:</strong> Breathe life into your UI with a
-            straightforward signal-based reactivity model. Watch as your UI updates magically as
-            your data changes.
-          </li>
-          <li>
-            <strong>🧩 Zero Dependencies:</strong> With Elemental.js, what you see is what you get -
-            no external dependencies, no bloated frameworks, just pure vanilla JavaScript wizardry.
-          </li>
-          <li>
-            <strong>🚀 Lightweight and Fast:</strong> Designed to be lean and mean. Enjoy the speed
-            without the bloat of larger frameworks.
-          </li>
-          <li>
-            <strong>📚 Easy to Learn:</strong> Jump right in, even if you're not a sorcerer of
-            JavaScript. Clear HTML-like syntax makes it a breeze.
-          </li>
-          <li>
-            <strong>Compatible:</strong> Works seamlessly with any JavaScript framework or library.
-            Use it alongside your favorite tools without a hitch.
-          </li>
-        </ul>
-      </section>
-      <section>
-        <h2>To Get Started with Elemental.js</h2>
-        <p>Add it through a <strong>CDN</strong>:</p>
-        <pre>
-          <code class="language-html">
-            &lt;script src="https://cdn.jsdelivr.net/npm/elemental-js@latest">&lt;/script>
-          </code>
-        </pre>
-        <p>
-          Or install via <strong>npm</strong> or whatever package manager you prefer, though I
-          really recommend you to try out <a href="https://bun.sh/"><strong>Bun</strong></a
-          >/<a href="https://deno.com/"><strong>Deno</strong></a
-          >:
-        </p>
-        <div class="tab-container">
-          <div class="tab-buttons">
-            <button class="btn active" type="menu" data-tab="bun">bun</button>
-            <button class="btn" type="menu" data-tab="npm">npm</button>
-            <button class="btn" type="menu" data-tab="pnpm">pnpm</button>
-            <button class="btn" type="menu" data-tab="yarn">yarn</button>
-          </div>
-          <div class="tab-content">
-            <pre data-tab="bun">
-              <code class="language-sh">
-                bun add elemental-js
-              </code>
-            </pre>
-            <pre data-tab="npm" hidden>
-              <code class="language-sh">
-                npm install elemental-js
-              </code>
-            </pre>
-            <pre data-tab="pnpm" hidden>
-              <code class="language-sh">
-                pnpm add elemental-js
-              </code>
-            </pre>
-            <pre data-tab="yarn" hidden>
-              <code class="language-sh">
-                yarn add elemental-js
-              </code>
-            </pre>
-          </div>
-        </div>
-      </section>
-      <section>
-        <h2>Counter Example</h2>
-        <pre>
-          <code class="language-js">
-            import { el, def } from 'elemental-js'
-  
-            function Counter() {
-              const count = def(0)
-              const increment = () => ++count.val
-
-              return el`button onclick=${increment}`(
-                'Count is: ', count
-              )
-            }
-  
-            l('#counter').append(Counter())
-          </code>
-          
-        </pre>
-        <p>
-          Check out more <a href="/examples" class="button">examples</a>, and embrace the power of
-          Elemental.js. Let the elements guide your path to creating enchanting user interfaces. Who
-          knew controlling the elements could be this easy?
-        </p>
-      </section>
-      <section>
-        <h2>Why</h2>
-        <p>
-          Elemental.js was born out of a desire for a lightweight, dependency-free JavaScript
-          library that could be used to write userscripts. Also for learning and fun.
-        </p>
-      </section>
-    </main>
-    <footer>
-      Made with <span>❤️</span> by
-      <a href="https://github.com/Stefanovicius" target="_blank">Adomas Stefanovičius</a>
-    </footer>
-  </body>
-</html>
