--- packages/site/vite.config.js
@@ -1,6 +1,0 @@
-import { resolve } from 'path'
-import { defineConfig } from 'vite'
-
-export default defineConfig({
-  root: resolve(__dirname, 'src')
-})
