+++ src/element/core.js
