+++ src/element/text.js
