+++ tests/children.test.js
@@ -1,6 +1,7 @@
 import { describe, it, expect } from 'vitest'
 import { handleChildren } from '../src/element/children'
-import { el, def } from '../src/main'
+import { handleTemplate } from '../src/element/core'
+import { createReactive } from '../src/reactive/core'
 
 describe('Handling children', () => {
   it('should append text nodes correctly', () => {
@@ -17,7 +18,7 @@ describe('Handling children', () => {
 
   it('should handle reactive children', async () => {
     const element = document.createElement('div')
-    const reactive = def('Initial')
+    const reactive = createReactive('Initial')
     const children = [reactive]
 
     handleChildren(element, children)
@@ -31,7 +32,7 @@ describe('Handling children', () => {
   it('should handle mixed children types', () => {
     const element = document.createElement('div')
     const child1 = document.createElement('span')
-    const reactive = def('Reactive')
+    const reactive = createReactive('Reactive')
     const children = [child1, reactive, 'Static']
 
     handleChildren(element, children)
@@ -44,8 +45,8 @@ describe('Handling children', () => {
   it('should remove extra children when updating', async () => {
     const element = document.createElement('div')
 
-    const [reactive, children] = def(['One', 'Two', 'Three'], (value) =>
-      value.map((text) => el`span`(text))
+    const [reactive, children] = createReactive(['One', 'Two', 'Three'], (value) =>
+      value.map((text) => handleTemplate`span`(text))
     )
 
     handleChildren(element, [children])
